Remove `Address` model, merge with `Location`
