+++ app/models/location.rb
@@ -1,7 +1,6 @@
 # frozen_string_literal: true
 class Location < ApplicationRecord # rubocop: disable Metrics/ClassLength
   EDIT_FIELDS = %w(
-    address_id
     booking_location_uid
     hidden
     hours
@@ -14,11 +13,16 @@ class Location < ApplicationRecord # rubocop: disable Metrics/ClassLength
     online_booking_twilio_number
     online_booking_enabled
     online_booking_reply_to
+    address_line_1,
+    address_line_2,
+    address_line_3,
+    town,
+    county,
+    postcode
   ).freeze
   TP_CALL_CENTRE_NUMBER = '+442037333495'
   ORGANISATIONS = %w(cas cita nicab).freeze
 
-  belongs_to :address, validate: true
   belongs_to :booking_location,
              foreign_key: :booking_location_uid,
              primary_key: :uid,
@@ -32,7 +36,6 @@ class Location < ApplicationRecord # rubocop: disable Metrics/ClassLength
   validates :uid, presence: true
   validates :organisation, presence: true, inclusion: ORGANISATIONS
   validates :title, presence: true
-  validates :address, presence: true
   validates :booking_location, presence: { if: ->(l) { l.phone.blank? } }
   validates :phone,
             presence: { if: ->(l) { l.booking_location.blank? } },
@@ -49,6 +52,10 @@ class Location < ApplicationRecord # rubocop: disable Metrics/ClassLength
             if: :online_booking_enabled?
   validates :hidden, inclusion: { in: [true], if: ->(l) { l.twilio_number.blank? } }
   validates :online_booking_enabled, inclusion: { in: [true, false] }
+  validates :postcode, :address_line_1, presence: true
+  validate :valid_uk_postcode
+
+  before_validation :set_point_from_postcode
 
   default_scope -> { order(:title) }
   scope :active, -> { where(hidden: false) }
@@ -71,15 +78,15 @@ class Location < ApplicationRecord # rubocop: disable Metrics/ClassLength
     end
 
     def externally_visible(include_hidden_locations:)
-      scope = includes(:address, :booking_location)
+      scope = includes(:booking_location)
       scope = scope.active unless include_hidden_locations
       scope
     end
 
     def with_visibility_flags(hidden_flags)
       where(hidden: hidden_flags)
-        .includes(:address, :editor, :booking_location)
-        .references(:address, :editor, :booking_location)
+        .includes(:editor, :booking_location)
+        .references(:editor, :booking_location)
     end
 
     def latest_for_twilio_number
@@ -112,8 +119,19 @@ class Location < ApplicationRecord # rubocop: disable Metrics/ClassLength
     booking_location.nil? && phone.present?
   end
 
+  def address_lines
+    [
+      address_line_1,
+      address_line_2,
+      address_line_3,
+      town,
+      county,
+      postcode
+    ].map(&:presence).compact
+  end
+
   def address_line
-    address.to_a.join(', ')
+    address_lines.compact
   end
 
   def slots
@@ -130,5 +148,28 @@ class Location < ApplicationRecord # rubocop: disable Metrics/ClassLength
 
   def operational?
     !cut_off?
+  end
+
+  def valid_uk_postcode
+    return if postcode.blank?
+
+    uk_postcode = UKPostcode.parse(postcode)
+    if uk_postcode.full_valid?
+      errors.add(:postcode, :geocoding_error) unless point.present?
+    else
+      errors.add(:postcode, :non_uk)
+    end
+  end
+
+  def set_point_from_postcode
+    return if postcode.blank? || point.present?
+
+    geocode = PostcodeGeocoder.new(postcode)
+    return unless geocode.valid?
+
+    self.point = {
+      type: 'Point',
+      coordinates: geocode.coordinates
+    }
   end
 end
