+++ app/views/locations/index.json.jbuilder
@@ -2,10 +2,10 @@ json.type 'FeatureCollection'
 json.features @locations do |location|
   json.type 'Feature'
   json.id location.uid
-  json.geometry location.address.point
+  json.geometry location.point
   json.properties do
     json.title location.title
-    json.address location.address.to_a.join("\n")
+    json.address location.address_lines.join("\n")
     json.booking_location_id location.booking_location&.uid.to_s
     json.phone location.phone.to_s
     json.hours location.hours.to_s
